+++ src/roiextractors/extractors/schnitzerextractor/extractsegmentationextractor.py
@@ -1,6 +1,7 @@
 """Extractor for reading the segmentation data that results from calls to EXTRACT."""
 from abc import ABC
 from pathlib import Path
+from typing import Optional
 
 import numpy as np
 from lazy_ops import DatasetView
@@ -34,7 +35,7 @@ class ExtractSegmentationExtractor(ABC):
         cls,
         file_path: PathType,
         sampling_frequency: float,
-        output_struct_name: str = "output",
+        output_struct_name: Optional[str] = None,
     ):
         """Abstract class that defines which extractor class to use for a given file.
         For newer versions of the EXTRACT algorithm, the extractor class redirects to
@@ -45,16 +46,26 @@ class ExtractSegmentationExtractor(ABC):
         ----------
         file_path: str
             The location of the folder containing the .mat file.
-        output_struct_name: str
-            The name of output struct in the .mat file, default is "output".
+        output_struct_name: str, optional
+            The name of output struct in the .mat file.
+            When unspecified, we check if any of the default values can be found in the file.
+            For newer version of extract, the default name is assumed to be "output".
+            For older versions the default is "extractAnalysisOutput". If none of them
+            can be found, it must be supplied.
         sampling_frequency: float
             The sampling frequency in units of Hz.
         """
         self = super().__new__(cls)
         self.file_path = file_path
-        self.output_struct_name = output_struct_name
         # Check if the file is a .mat file
         cls._assert_file_is_mat(self)
+
+        if output_struct_name is None:
+            self.output_struct_name = cls._get_default_output_struct_name_from_file(self)
+        else:
+            # Check that user-given 'output_struct_name' is in the file
+            self.output_struct_name = output_struct_name
+            cls._assert_output_struct_name_is_in_file(self)
 
         # Check the version of the .mat file
         if cls._check_extract_file_version(self):
@@ -62,11 +73,14 @@ class ExtractSegmentationExtractor(ABC):
             return NewExtractSegmentationExtractor(
                 file_path=file_path,
                 sampling_frequency=sampling_frequency,
-                output_struct_name=output_struct_name,
+                output_struct_name=self.output_struct_name,
             )
 
         # For older versions of the .mat file, use the legacy extractor
-        return LegacyExtractSegmentationExtractor(file_path=file_path)
+        return LegacyExtractSegmentationExtractor(
+            file_path=file_path,
+            output_struct_name=self.output_struct_name,
+        )
 
     def _assert_file_is_mat(self):
         """Check that the file exists and is a .mat file."""
@@ -74,19 +88,39 @@ class ExtractSegmentationExtractor(ABC):
         assert file_path.exists(), f"File {file_path} does not exist."
         assert file_path.suffix == ".mat", f"File {file_path} must be a .mat file."
 
+    def _get_default_output_struct_name_from_file(self):
+        """Return the default value for 'output_struct_name' when it is unspecified.
+        For newer version of extract, the default name is assumed to be "output".
+        For older versions the default is "extractAnalysisOutput".
+        If none of them is found, raise an error that 'output_struct_name' must be supplied."""
+        newer_default_output_struct_name = "output"
+        legacy_default_output_struct_name = "extractAnalysisOutput"
+        with h5py.File(name=self.file_path, mode="r") as mat_file:
+            if newer_default_output_struct_name in mat_file.keys():
+                return newer_default_output_struct_name
+            elif legacy_default_output_struct_name in mat_file.keys():
+                return legacy_default_output_struct_name
+            else:
+                raise AssertionError("The 'output_struct_name' must be supplied.")
+
+    def _assert_output_struct_name_is_in_file(self):
+        """Check that 'output_struct_name' is in the file, raises an error if not."""
+        with h5py.File(name=self.file_path, mode="r") as mat_file:
+            assert (
+                self.output_struct_name in mat_file
+            ), f"Output struct name '{self.output_struct_name}' not found in file."
+
     def _check_extract_file_version(self) -> bool:
         """Check the version of the extract file.
         If the file was created with a newer version of the EXTRACT algorithm, the
         function will return True, otherwise it will return False."""
         with h5py.File(name=self.file_path, mode="r") as mat_file:
-            if self.output_struct_name not in mat_file:
-                return False
             dataset_version = mat_file[self.output_struct_name]["info"]["version"][:]
             dataset_version = np.ravel(dataset_version)
             # dataset_version is an HDF5 dataset of encoded characters
             version_name = _decode_h5py_array(dataset_version)
 
-            return version.Version(version_name) >= version.Version("1.1.0")
+            return version.Version(version_name) >= version.Version("1.0.0")
 
 
 class NewExtractSegmentationExtractor(SegmentationExtractor):
@@ -180,7 +214,7 @@ class NewExtractSegmentationExtractor(SegmentationExtractor):
         return DatasetView(self._output_struct["spatial_weights"]).lazy_transpose()
 
     def _trace_extractor_read(self) -> DatasetView:
-        """Returns the traces with a shape of number of ROIs and number of frames."""
+        """Returns the traces with a shape of number of frames and number of ROIs."""
         return DatasetView(self._output_struct["temporal_weights"]).lazy_transpose()
 
     def get_accepted_list(self) -> list:
@@ -193,7 +227,7 @@ class NewExtractSegmentationExtractor(SegmentationExtractor):
         accepted_list: list
             List of accepted ROIs
         """
-        return [roi for roi in self.get_roi_ids() if np.any(self._image_masks[roi])]
+        return [roi for roi in self.get_roi_ids() if np.any(self._image_masks[..., roi])]
 
     def get_rejected_list(self) -> list:
         """
@@ -234,7 +268,7 @@ class NewExtractSegmentationExtractor(SegmentationExtractor):
     def get_images_dict(self):
         """
         Returns a dictionary with key, values representing different types of Images
-        used in segmentation.
+        used in segmentation. The shape of images is height and width.
         Returns
         -------
         images_dict: dict
@@ -242,9 +276,9 @@ class NewExtractSegmentationExtractor(SegmentationExtractor):
         """
         images_dict = super().get_images_dict()
         images_dict.update(
-            summary_image=self._info_struct["summary_image"][:],
-            f_per_pixel=self._info_struct["F_per_pixel"][:],
-            max_image=self._info_struct["max_image"][:],
+            summary_image=self._info_struct["summary_image"][:].T,
+            f_per_pixel=self._info_struct["F_per_pixel"][:].T,
+            max_image=self._info_struct["max_image"][:].T,
         )
 
         return images_dict
@@ -263,16 +297,24 @@ class LegacyExtractSegmentationExtractor(SegmentationExtractor):
     mode = "file"
     installation_mesg = "To use extract install h5py: \n\n pip install h5py \n\n"  # error message when not installed
 
-    def __init__(self, file_path: PathType):
+    def __init__(
+        self,
+        file_path: PathType,
+        output_struct_name: str = "extractAnalysisOutput",
+    ):
         """
         Parameters
         ----------
         file_path: str
             The location of the folder containing dataset.mat file.
+        output_struct_name: str, optional
+            The user has control over the names of the variables that return from `extraction(images, config)`.
+            When unspecified, the default is 'extractAnalysisOutput'.
         """
         super().__init__()
         self.file_path = file_path
-        self._dataset_file, self._group0 = self._file_extractor_read()
+        self._dataset_file = self._file_extractor_read()
+        self.output_struct_name = output_struct_name
         self._image_masks = self._image_mask_extractor_read()
         self._roi_response_raw = self._trace_extractor_read()
         self._raw_movie_file_location = self._raw_datafile_read()
@@ -283,27 +325,24 @@ class LegacyExtractSegmentationExtractor(SegmentationExtractor):
         self._dataset_file.close()
 
     def _file_extractor_read(self):
-        f = h5py.File(self.file_path, "r")
-        _group0_temp = list(f.keys())
-        _group0 = [a for a in _group0_temp if "#" not in a]
-        return f, _group0
+        return h5py.File(self.file_path, "r")
 
     def _image_mask_extractor_read(self):
-        return self._dataset_file[self._group0[0]]["filters"][:].transpose([1, 2, 0])
+        return self._dataset_file[self.output_struct_name]["filters"][:].transpose([1, 2, 0])
 
     def _trace_extractor_read(self):
-        return self._dataset_file[self._group0[0]]["traces"]
+        return self._dataset_file[self.output_struct_name]["traces"]
 
     def _tot_exptime_extractor_read(self):
-        return self._dataset_file[self._group0[0]]["time"]["totalTime"][0][0]
+        return self._dataset_file[self.output_struct_name]["time"]["totalTime"][0][0]
 
     def _summary_image_read(self):
-        summary_image = self._dataset_file[self._group0[0]]["info"]["summary_image"]
+        summary_image = self._dataset_file[self.output_struct_name]["info"]["summary_image"]
         return np.array(summary_image)
 
     def _raw_datafile_read(self):
-        if self._dataset_file[self._group0[0]].get("file"):
-            charlist = [chr(i) for i in np.squeeze(self._dataset_file[self._group0[0]]["file"][:])]
+        if self._dataset_file[self.output_struct_name].get("file"):
+            charlist = [chr(i) for i in np.squeeze(self._dataset_file[self.output_struct_name]["file"][:])]
             return "".join(charlist)
 
     def get_accepted_list(self):